+++ lib/ui/alignment.dart
@@ -1,0 +1,4 @@
+
+import 'package:flutter/material.dart';
+
+extension AlignX<T extends Widget> on T {}